tests/wpt/tests/html/canvas/offscreen/fill-and-stroke-styles/2d.gradient.object.update.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.gradient.object.update</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.gradient.object.update</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  var g = ctx.createLinearGradient(-100, 0, 200, 0);
  g.addColorStop(0, '#f00');
  g.addColorStop(1, '#f00');
  ctx.fillStyle = g;
  ctx.fillRect(0, 0, 100, 50);
  g.addColorStop(0.1, '#0f0');
  g.addColorStop(0.9, '#0f0');
  ctx.fillRect(0, 0, 100, 50);
  _assertPixelApprox(canvas, 50,25, 0,255,0,255, 2);
  t.done();

});
</script>
